feat: :sparkles: organizado arquivos de migração, além das fuhnções de sync do erp
