--- api/alembic/versions/ae0abbc1bdda_11_mercado_produto_promocao.py
@@ -1,65 +1,0 @@
-"""11_mercado_produto_promocao
-
-Revision ID: ae0abbc1bdda
-Revises: c01e4be735cd
-Create Date: 2024-04-26 14:38:11.846234
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import fastapi_storages
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = 'ae0abbc1bdda'
-down_revision: Union[str, None] = 'c01e4be735cd'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-storage = fastapi_storages.FileSystemStorage(path="./media")
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('mercado_promocao',
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('produto', postgresql.JSONB(astext_type=sa.Text()), nullable=False),
-    sa.Column('data_inicial', sa.DateTime(), nullable=False),
-    sa.Column('data_final', sa.DateTime(), nullable=False),
-    sa.Column('percentual_desconto', sa.Float(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=False),
-    sa.Column('deleted', sa.Boolean(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('mercado_produto',
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('mercado_id', sa.UUID(), nullable=True),
-    sa.Column('nome', sa.String(length=255), nullable=True),
-    sa.Column('marca', sa.String(length=255), nullable=True),
-    sa.Column('data_validade', sa.DateTime(), nullable=True),
-    sa.Column('ncm_produto', sa.String(length=10), nullable=True),
-    sa.Column('gtin_produto', sa.String(length=14), nullable=True),
-    sa.Column('mpn_produto', sa.String(length=30), nullable=True),
-    sa.Column('id_produto_erp', sa.String(), nullable=True),
-    sa.Column('descricao', sa.String(length=500), nullable=True),
-    sa.Column('preco', sa.Float(), nullable=True),
-    sa.Column('preco_promocional', sa.Float(), nullable=True),
-    sa.Column('imagem', fastapi_storages.integrations.sqlalchemy.ImageType(storage=storage), nullable=True),
-    sa.Column('codigo_produto', sa.String(length=30), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=False),
-    sa.Column('deleted', sa.Boolean(), nullable=False),
-    sa.ForeignKeyConstraint(['mercado_id'], ['mercado_mercado.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('mercado_produto')
-    op.drop_table('mercado_promocao')
-    # ### end Alembic commands ###
